GERONIMO-5624, GERONIMO-5577 Use proposed modifications in tomcat securiity contracts to avoid wrapping servlet context and ServletRegistration.Dynamic
git-svn-id: https://svn.apache.org/repos/asf/geronimo/server/trunk@1001836 13f79535-47bb-0310-9956-ffa450edef68

+++ plugins/tomcat/geronimo-tomcat7/src/main/java/org/apache/geronimo/tomcat/stats/ModuleStats.java
@@ -120,9 +120,9 @@ public class ModuleStats {
             stats.setSessionAliveTime(maxActive, -1, sessionMaxAliveTime,
                     sessionAverageAliveTime * maxActive);
             stats.setRejectedSessionCount(rejectedSessions);
-            stats.setExpiredSessionCount(expiredSessions);
+            stats.setExpiredSessionCount((int)expiredSessions);
             stats.setActiveSessionCount(activeSessions);
-            stats.setSessionCount(sessionCounter);
+            stats.setSessionCount((int)sessionCounter);
 
         } catch (Exception ex) {
             log.error("Error getting attribute " + mgrName + " " + ex.toString());
